# Enable the Information Schema Connector to load data lazily