… presto-main/src/main/java/io/prestosql/connector/informationschema/InformationSchemaTableHandle.java
… presto-main/src/test/java/io/prestosql/connector/MockConnectorFactory.java
… presto-main/src/test/java/io/prestosql/metadata/TestInformationSchemaMetadata.java
… presto-main/src/test/java/io/prestosql/metadata/TestInformationSchemaTableHandle.java
… presto-tests/src/test/java/io/prestosql/connector/informationschema/BenchmarkInformationSchema.java
… presto-tests/src/test/java/io/prestosql/tests/TestInformationSchemaConnector.java