This commit changes both metadata side and read side of the Information Schema Connector:
1. Metadata side: this commit removes the computation logic in `InformationSchemaMetadata::applyFilter` and delegates it to `InformationSchemaPageSource`. The change reduces the time of query planning, so that it doesn't bottleneck the query execution.
2. Read side: this commit makes `InformationSchemaPageSource` be able to build pages on the fly when `getNextPage` is called. Before this change, all the pages are built beforehand before downstream tasks can start execution.

This commit also adds a more comprehensive set of unit tests for the Information Schema Connector.

This lazy-dataloading feature brings the following benefits:
1. It decreases the latency of first query result.
2. It benefits queries with a limit clause, e.g., `SELECT * FROM hive.information_schema.columns LIMIT 1`.
3. For queries touching a lot of tables (e.g., `SELECT * from hive.information_schema.columns`), it's highly possible that some tables have invalid metadata, which eventually fails the query. The new approach passes Pages to downstream tasks as soon as they are ready. In this way, failures can be detected sooner.

Running BenchmarkInformationSchema.queryInformationSchema
```
Before:
Benchmark                                          (columnsCount)        (queryId)  (schemasCount)  (tablesCount)  Mode  Cnt     Score     Error  Units
BenchmarkInformationSchema.queryInformationSchema             100        FULL_SCAN             200            200  avgt   10  2838.890 ± 349.474  ms/op
BenchmarkInformationSchema.queryInformationSchema             100   LIKE_PREDICATE             200            200  avgt   10   114.690 ±  21.006  ms/op
BenchmarkInformationSchema.queryInformationSchema             100  MIXED_PREDICATE             200            200  avgt   10    69.226 ±  14.637  ms/op
BenchmarkInformationSchema.queryInformationSchema             100       LIMIT_SCAN             200            200  avgt   10  3071.904 ± 345.259  ms/op
```


```
After:
Benchmark                                          (columnsCount)        (queryId)  (schemasCount)  (tablesCount)  Mode  Cnt     Score     Error  Units
BenchmarkInformationSchema.queryInformationSchema             100        FULL_SCAN             200            200  avgt   10  2286.313 ± 109.109  ms/op
BenchmarkInformationSchema.queryInformationSchema             100   LIKE_PREDICATE             200            200  avgt   10    41.420 ±  10.933  ms/op
BenchmarkInformationSchema.queryInformationSchema             100  MIXED_PREDICATE             200            200  avgt   10    70.408 ±  13.001  ms/op
BenchmarkInformationSchema.queryInformationSchema             100       LIMIT_SCAN             200            200  avgt   10    45.527 ±   8.346  ms/op
```
#1046 #998