@linxingyuan1102 2019-07-13
Tests done:
1. All tests in `TestInformationSchemaConnector` passed. They are written to test the correctness of the new implementation.
2. I tested manually against a Hive catalog with `select * from hive.information_schema.columns`. The new implementation does achieve the benefits mentioned above.

@kokosing Could you take a look?

cc: @wagnermarkd @phd3 @electrum @findepi