@kokosing 2019-07-15
Can you please give some example numbers. Like having # of schemas, tables, columns in ? connector and queries like *** execution time changed from `x ms` to `s ms`.