@linxingyuan1102 2019-07-16
Thanks for the review. @kokosing 

> Can you please give some example numbers. Like having # of schemas, tables, columns in ? connector and queries like *** execution time changed from `x ms` to `s ms`.

I added a query `SELECT column_name FROM information_schema.columns LIMIT 100` in `BenchmarkInformationSchema`. The execution time of this query changed from `2819.921 ± 352.867  ms/op` to `51.422 ±   7.855  ms/op`