@kokosing 2019-07-30
@linxingyuan1102 besides the comments there is one more thing to add. 

Previously code was optimized for case like `SELECT * FROM information_schema.columns WHERE table_name LIKE 'some_table' AND table_schema LIKE 'some_schema'`. Notice that `_` is like `.` in regex. So query matches to tables like: `some_schema.some_table` or `someXschema.someXtable`. This is common bug SQL notebooks (see https://sourceforge.net/p/squirrel-sql/bugs/1359/). The idea was to enumerate schemas and tables and then in add exchanges optimizer filter only matching schemas and tables. You are still doing that for schemas (which is good), but you don't do it for tables. Which could lead to a regression. Hence prefixes were used because prefix stores information schema table pair.

I think that besides correctness test and benchmark (based on mock) we need to do either:
 - do benchmarks on real connector when querying schemas, tables or columns takes non trivial 
 - or do test on hive connector and monitor how many times metastore was queried and make sure that we do not call them more than is needed (or was before).