@linxingyuan1102 2019-07-30
@kokosing Thanks for the review.
Let's discuss more about the regression cases here.
As an example, consider a hive catalog with `M` schemas and each schema has `N` tables averagely.

For tables `schemata`, `roles`, `applicable_roles` and `enabled_roles`: there should be no regression and each query should incur `O(1)` metastore calls.

For tables `tables` and `views`: there should be no regression and each query should incur at most `O(M)` metastore calls.

For tables `columns`: there are some cases which lead to a regression:
1. For queries like `SELECT * FROM information_schema.columns WHERE table_name LIKE 'some_table' AND table_schema LIKE 'some_schema'`, the number of metastore calls regresses from `O(1)` to `O(N)`. (Assume the LIKE predicate returns a `O(1)` result)
2. For queries like `SELECT * FROM information_schema.columns WHERE table_name LIKE 'some_table'`, the number of metastore calls regresses from `O(M)` to `O(M*N)`. (Also assume the LIKE predicate returns a `O(1)` result)

I think we can make some changes to the Hive Connector to avoid such regression in terms of number of metastore calls. In its current implementation, Hive Connector only uses [ThriftHiveMetastore.Iface::get_table](https://hive.apache.org/javadocs/r1.2.2/api/org/apache/hadoop/hive/metastore/api/ThriftHiveMetastore.Iface.html#get_table(java.lang.String,%20java.lang.String)) to get the column information. So it fires a metastore call table by table. However, there is a batch version of this method [ThriftHiveMetastore.Iface::get_table_objects_by_name](https://hive.apache.org/javadocs/r1.2.2/api/org/apache/hadoop/hive/metastore/api/ThriftHiveMetastore.Iface.html#get_table_objects_by_name(java.lang.String,%20java.util.List)). I think we can utilize it and make some changes in `HiveMetadata::listTableColumns`.  What's your opinion?

Of course, it definitely helps to do some benchmark work. We can discuss more about this.