@kokosing 2019-07-31
> I think we can make some changes to the Hive Connector to avoid such regression in terms of number of metastore calls.

That would also require a change in the connector interface. That would be a major change that sounds like it would beneficial in general, however I would prefer no to block this PR on this. Let's create an issue for that.

Maybe to not to regress today, we could restore "previous" behavior only for `columns` table and only if list of schemas in `TableHandle` is no bigger than 10 and table names absent in `TableHandle`.

Also notice that even if we add `getTables` (batch version of `getTable`) to connector interface it is still beneficial to filter tables early as then list of table names and then the number of processed returned tables will be smaller.