@linxingyuan1102 2019-07-31
@kokosing 
> That would also require a change in the connector interface. That would be a major change that sounds like it would beneficial in general, however I would prefer no to block this PR on this. Let's create an issue for that.

We can also limit the change within the Hive connector, i.e., changing `HiveMetadata::listTableColumns` directly. Further, it requires changes in `SemiTransactionalHiveMetastore`, `BridgingHiveMetastore`, `ThriftHiveMetastore`, `ThriftHiveMetastoreClient`, etc. I agree that we can work on this as another issue.

> Maybe to not to regress today, we could restore "previous" behavior only for `columns` table and only if list of schemas in `TableHandle` is no bigger than 10 and table names absent in `TableHandle`.

This is doable. We can add a field `Optional<Set<QualifiedTablePrefix>> prefixes` in `InformationSchemaTableHandle`. Use it when it is present and it's a `columns` table.

> Also notice that even if we add `getTables` (batch version of `getTable`) to connector interface it is still beneficial to filter tables early as then list of table names and then the number of processed returned tables will be smaller.

You're right. In the future if there is a way to pass the predicate function (in other forms) to the page source, regression will be totally avoided.